requirements/requirements.txt
numpy==1.23.5
pandas==1.5.3
pydantic==1.10.13
strictyaml==1.6.2
ruamel.yaml==0.17.32
joblib==1.3.2
scikit-learn==1.2.2
tensorflow==2.14.0
Pillow==9.4.0
opencv-python==4.8.0.76
matplotlib==3.7.1
wget==3.2
keras==2.14.0
mlflow==2.6.0